requirements.txt
cryptography
PyMySQL
python-dotenv
pydantic
rich
mysql
mysql-connector-python
cryptography
mysql.connector